ffrequirements.txt
asgiref==3.2.3
astroid==2.3.3
Babel==2.8.0
colorama==0.4.3
Django==3.0.3
django-crispy-forms==1.9.0
django-grappelli==2.14.1
django-jet==1.0.8
django-model-utils==4.0.0
django-notifications-hq==1.6.0
django-phonenumber-field==4.0.0
django-tables2==2.2.1
isort==4.3.21
jsonfield==3.1.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
phonenumbers==8.11.4
Pillow==7.0.0
pylint==2.4.4
pytz==2019.3
six==1.14.0
sqlparse==0.3.1
swapper==1.1.2.post1
typed-ast==1.4.1
wrapt==1.11.2
